112 separate CodeGenerator tests
